Removed the reference to the raising of the BadGateway if the destination host and the current host are not the same.
This is referenced in the rfc2518 as:

502 (Bad Gateway) - This may occur when the destination is on another
   server and the destination server refuses to accept the resource.

as this working makes it seem that it is context specific and optional. my current setup makes it so that this is always fails.  I could have added a feature toggle and i may still in the future but for now i am going to just remove this check.